backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Validation and settings
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# WebSocket support (included in uvicorn[standard])
websockets==12.0

# Async file operations
aiofiles==23.2.1

# YAML for Claude config files
pyyaml==6.0.1
